// src/types/clientsTypes.ts
export interface Client {
  id: number;
  name: string;
  description: string;
  industry: string;
  contactPerson: string;
  email: string;
  phoneNumber: string;
  revenueGenerated: number;
  status: string;
  website: string;
  lastInteractionDate: string;
  notes: string;
  preferredCommunicationMethod: string;
  clientType: string;
}